add: user (employer and candidate)registration

+++ src/main/java/com/project/miniJobPortal/repository/SkillRepository.java
@@ -3,5 +3,5 @@ package com.project.miniJobPortal.repository;
 import com.project.miniJobPortal.entity.Skill;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SkillRepository extends JpaRepository<Skill, Integer> {
+public interface SkillRepository extends JpaRepository<Skill, Long> {
 }
